+++ src/Provider/Hook_Event_Provider.php
@@ -6,6 +6,7 @@ namespace Rarst\Laps\Provider;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
 use Rarst\Laps\Event\Core_Events;
+use Rarst\Laps\Event\Extension_Events;
 use Rarst\Laps\Event\Genesis_Events;
 use Rarst\Laps\Event\Hybrid_Events;
 use Rarst\Laps\Event\THA_Events;
@@ -25,6 +26,7 @@ class Hook_Event_Provider implements ServiceProviderInterface {
 		$pimple['hook.events'] = function (): array {
 			return [
 				'core'        => new Core_Events(),
+				'extension'   => new Extension_Events(),
 				'genesis'     => new Genesis_Events(),
 				'hybrid'      => new Hybrid_Events(),
 				'tha'         => new THA_Events(),